Update of mainwindow
The main window take care now of the connection to the database, as well a the creation of the document model and its connection with its specific view, and trigger the creation of a c_add_document GUI via the menu for the creation of a new document.
The insertion of the new document in the model is not handle at the moment.